fix: change jitter code to avoid bias
The existing jitter code used a random sampling from 0-newDelay, but this
introduces a significant bias. Because it then applied the min and max, it
wasn't 50% but just 60% in testing, but it still means that when a human says
"I want to wait 5s" it would average 3s wait, which is a mismatched expectation.

I also found that the StopChannel test was flaky, because the clock could tick
and a new event be generated from the clock while the stop channel was being closed.
1us wasn't reliable, 10us never failed in fast testing.