fix: handle bad exponents to ExpBackoff
With the previous implementation, it always checked if you would go
above MaxDelay, but only checked if you were below MinDelay if you used
jitter. In practice, however, we've passed in bad arguments, and then
the delays to terrible things. Ideally it would be a user-error (compile
time error), but for now we can at least make the system resilient and
not end up with 'exponential backoff' actually going to 0 delay.